+++ test/test_mongo_controller.py
@@ -9,6 +9,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo
+import layabase.testing
 
 # Use debug logging to ensure that debug logging statements have no impact
 logging.basicConfig(
@@ -39,7 +40,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 @pytest.fixture