+++ test/test_mongo_controller_audit_enum.py
@@ -4,6 +4,7 @@ import re
 import pytest
 
 from layabase import database, database_mongo
+import layabase.testing
 
 
 class EnumTest(enum.Enum):
@@ -30,7 +31,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock?ssl=True", _create_models, replicaSet="globaldb")
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_post_with_enum_is_valid(db):