+++ test/test_mongo_controller_audit_int.py
@@ -3,6 +3,7 @@ import re
 import pytest
 
 from layabase import database, database_mongo
+import layabase.testing
 
 
 class TestIntController(database.CRUDController):
@@ -23,7 +24,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock?ssl=True", _create_models, replicaSet="globaldb")
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_int_revision_is_not_reset_after_delete(db):