+++ test/test_mongo_controller_audit_versioned_rollback_forbidden.py
@@ -5,6 +5,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo, versioning_mongo
+import layabase.testing
 
 
 class EnumTest(enum.Enum):
@@ -40,7 +41,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock?ssl=True", _create_models, replicaSet="globaldb")
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_rollback_validation_custom(db):