+++ test/test_mongo_controller_id.py
@@ -1,7 +1,9 @@
 import pytest
 
-from layabase import database, database_mongo
 from layaberr import ValidationFailed
+
+from layabase import database, database_mongo
+import layabase.testing
 
 
 class TestIdController(database.CRUDController):
@@ -21,7 +23,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_post_id_is_valid(db):