+++ test/test_mongo_controller_index.py
@@ -5,6 +5,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo
+import layabase.testing
 
 
 class TestIndexController(database.CRUDController):
@@ -29,7 +30,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_post_twice_with_unique_index_is_invalid(db):