+++ test/test_mongo_controller_int_float.py
@@ -2,6 +2,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo
+import layabase.testing
 
 
 class TestIntAndFloatController(database.CRUDController):
@@ -24,7 +25,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_post_int_str_in_int_column(db):