+++ test/test_mongo_controller_strict.py
@@ -2,6 +2,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo
+import layabase.testing
 
 
 class TestStrictController(database.CRUDController):
@@ -28,7 +29,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 def test_post_with_unknown_field_is_invalid(db):