+++ test/test_mongo_controller_unique_non_primary.py
@@ -4,6 +4,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo, versioning_mongo
+import layabase.testing
 
 
 class EnumTest(enum.Enum):
@@ -67,7 +68,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 class DateTimeModuleMock: