+++ test/test_mongo_controller_versionned.py
@@ -6,6 +6,7 @@ import pytest
 from layaberr import ValidationFailed
 
 from layabase import database, database_mongo, versioning_mongo
+import layabase.testing
 
 
 class EnumTest(enum.Enum):
@@ -41,7 +42,7 @@ def _create_models(base):
 def db():
     _db = database.load("mongomock", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 @pytest.fixture