+++ test/test_sqlalchemy_controller_query_required.py
@@ -5,6 +5,7 @@ import flask_restplus
 from layaberr import ValidationFailed
 
 from layabase import database, database_sqlalchemy
+import layabase.testing
 
 
 class TestRequiredController(database.CRUDController):
@@ -30,7 +31,7 @@ def _create_models(base):
 def db():
     _db = database.load("sqlite:///:memory:", _create_models)
     yield _db
-    database.reset(_db)
+    layabase.testing.reset(_db)
 
 
 @pytest.fixture